mac-sender/pyproject.toml
[build-system]
requires = ["setuptools>=61"]
build-backend = "setuptools.build_meta"

[project]
name = "keybridgebt-mac"
version = "2.0.0"
description = "keybridgeBT macOS sender — forward Mac keyboard/trackpad over Wi-Fi TCP"
requires-python = ">=3.11"
dependencies = [
    "pyobjc-framework-Quartz>=10.0",
    "PyNaCl>=1.5.0",
    "rumps>=0.4.0",
    "keyring>=25.0",
    "qrcode[pil]>=7.4",
    "Pillow>=10.0",
    "pyyaml>=6.0",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["keybridgebt_mac*"]
